Separate RelatedPage extensios into a separate file
